test(core): split core-extra into focused suites
Replace src/core/core-extra.test.ts with dedicated files:
- vDict.test.ts
- serialization.test.ts
- mobject.test.ts
- group.test.ts
- vgroup.test.ts
- vmobject.test.ts

Also update vmobject-coverage docs to reference vmobject.test.ts.
